LayoutTests/editing/pasteboard/ios/dom-paste-same-origin-expected.txt

Click here to copy
Click here to copy
Click here to copy
Verifies that programmatic paste is allowed when copied data is from the same origin. To manually test, tap the text on the bottom to programmatically copy, and then tap the editable area and check that the text is pasted twice.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.queryCommandSupported('Paste') is true
PASS document.queryCommandEnabled('Paste') is true
PASS event.clipboardData.getData('text/plain') is "Click here to copy"
PASS document.execCommand('Paste') is true
PASS event.clipboardData.getData('text/plain') is "Click here to copy"
PASS document.execCommand('Paste') is true
PASS editor.textContent is "Click here to copyClick here to copy"
PASS successfullyParsed is true

TEST COMPLETE
